Refactor package for demo
