+++ app/src/main/res/layout/activity_info.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.InfoActivity">
+    tools:context=".view.InfoActivity">
 
     <androidx.appcompat.widget.AppCompatImageView
         android:id="@+id/iv_info_image"
