+++ app/src/main/res/layout/activity_item_list.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.ItemListActivity">
+    tools:context=".view.ItemListActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerview_item_list"
