+++ app/src/main/res/layout/activity_launch.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.LaunchActivity">
+    tools:context=".view.LaunchActivity">
 
     <androidx.appcompat.widget.AppCompatImageView
         android:id="@+id/iv_launch_image"
