+++ app/src/main/res/layout/activity_main.xml
@@ -6,13 +6,13 @@
 
     <data>
         <variable name="viewModel"
-            type="com.kuanhsien.app.sample.android_mvvm_demo.ui.MainViewModel"/>
+            type="com.kuanhsien.app.sample.android_mvvm_demo.viewmodel.MainViewModel"/>
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.MainActivity">
+        tools:context=".view.MainActivity">
 
         <androidx.appcompat.widget.AppCompatImageView
             android:id="@+id/iv_cover_image"
